+++ src/commands/PRIVMSG.cpp
@@ -20,7 +20,7 @@ std::string PRIVMSG(std::vector<std::string> args,
   std::string target_nickname_or_channel = args[0];
   std::string message;
 
-  for (int i = 1; i < args.size(); i++) {
+  for (size_t i = 1; i < args.size(); i++) {
     if (i != 1)
       message.append(" ");
     message.append(args[i]);
